docker: adapt to latest python:3.8-slim-buster
When building the image with latest `python:3.8-slim-buster`, the build
was bringing over Python 2 based system packages instead of Python 3
based ones. This commit fixes the problem and makes the build a bit
slimmer and faster.

(However, Python 3.7 system packages are still being brought, which is
inconsistent with Python 3.8 version that the image provides. An upgrade
to Python 3.9 or a replacement of the base image may be wanted soon.)